# cuda : deterministic and faster k_copy_src1_to_contiguous