Upstreamed from ikawrakow/ik_llama.cpp#283 and ikawrakow/ik_llama.cpp#313