@CISC 2025-04-05
A quick benchmark after synchronization changes to show that it's still faster:

| Model             | CPU             | GPU        | n_ubatch |     test |        t/s master |     t/s deterministic_k_copy_src1 |
| :---------------- | :-------------- | :--------- | -------: | -------: | ----------------: | ------------------: |
| bailingmoe 16B Q8_0 | Core i7-9700K   | RTX 3090Ti |     2048 |   pp1024 |   3230.35 ± 22.55 |             3962.19 ± 37.60 |
| bailingmoe 16B Q8_0 | Core i7-9700K   | RTX 3090Ti |     2048 |   pp2048 |   3912.80 ± 24.77 |             4992.81 ± 43.04 |
| bailingmoe 16B Q8_0 | Core i7-9700K   | RTX 3090Ti |     2048 |   pp4096 |   3685.56 ± 21.49 |             4631.87 ± 26.74 |