@CISC 2025-04-05
> @CISC Please preserve the authorship of these patches as it was already suggested ([#1174 (comment)](https://github.com/ggml-org/ggml/pull/1174#issuecomment-2775253823)).

It feels wrong to fake authorship of the actual commits. Attribution in the PR itself (which can not be a direct upstream of the code anyway) should be sufficient?

Anyway, maybe it's better to ask @ikawrakow directly what his feelings about this is?