@ikawrakow 2025-04-05
> The code on master is deterministic in terms of results. What is not deterministic is how the work is distributed to CUDA blocks but I don't think that matters.

It is not my observation. I get a different result each time I run the code on `llama.cpp` master (on the same hardware, using the same parameters). The reason for this is that `src1` rows end up in a different order in the contiguous copy, which then influence the result as floating point operations are not associative. It is not wrong, but it is still not deterministic. If in doubt, run a few perplexity calculations and observe how the values are different in each run on master, but are always the same with this version of the kernel.