@CISC 2025-04-05
> In general I don't mind people copying my code, else I wouldn't be publishing it under a MIT license. But `llama.cpp` is special in this regard, so I think it is best to do the right thing. But as IANAL, and as I would like to avoid that someone comes along and starts teaching me that because of the blah blah convention blah blah blah, I leave it up to you and the project owner to determine what the right thing is.

What address would you like attributed, actual or the one in AUTHORS?