@ikawrakow 2025-04-06
> I'm reopening a new PR with co-authorship, the address in AUTHORS seems to link correctly.

Interesting approach. Because at some point in the past I contributed to this repository, from there on any original work I produce can be freely copied into here, and the attribution is covered by me being included in a list of a few hundred authors with a no-reply address?