@CISC 2025-04-06
> Interesting approach. Because at some point in the past I contributed to this repository, from there on any original work I produce can be freely copied into here, and the attribution is covered by me being included in a list of a few hundred authors with a no-reply address?

Only work published under a compatible license of course. The AUTHORS list is generated from git history, which is the only manageable way in a project with so many contributors I guess.

Anyway, I'm more than happy to make any additional attributions you wish, it is your work, and I'm very grateful for it!